27-ый commit 15-ый спринт, домашняя работа
