--- app/src/main/java/com/example/playlistmaker/ItunesAPI.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ItunesAPI {
-    @GET("/search?entity=song")
-    fun search(@Query("term") text: String): Call<SearchResponse>
-}
-
