+++ app/src/main/java/com/example/playlistmaker/MediaLibrary/MediaLibraryActivity.kt
@@ -1,0 +1,13 @@
+package com.example.playlistmaker.MediaLibrary
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.playlistmaker.R
+
+class MediaLibraryActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_medialibrary)
+    }
+
+}
