--- app/src/main/java/com/example/playlistmaker/MediaLibraryActivity.kt
@@ -1,12 +1,0 @@
-package com.example.playlistmaker
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MediaLibraryActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_medialibrary)
-    }
-
-}
