+++ app/src/main/java/com/example/playlistmaker/Player/domain/PlayerConst.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.Player.domain
+
+object PlayerConst {
+    const val TRACK_INFO = "track_info"
+    const val DEFAULT_DURATION = "00:00"
+    const val DURATION_REFRESH_DELAY = 400L
+}
