+++ app/src/main/java/com/example/playlistmaker/Search/data/dto/NetworkResponse.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.Search.data.dto
+
+open class NetworkResponse {
+    var resultCode = 0
+}
