+++ app/src/main/java/com/example/playlistmaker/Search/data/dto/SearchRequest.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.Search.data.dto
+
+data class SearchRequest(val expression: String)
