+++ app/src/main/java/com/example/playlistmaker/Search/data/dto/SearchResponse.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.Search.data.dto
+
+data class SearchResponse(
+    val resultCount: Int,
+    val results: List<TrackDto>
+) : NetworkResponse()
+
