+++ app/src/main/java/com/example/playlistmaker/Search/data/repository/TrackListClickListenerInterface.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.Search.data.repository
+
+import com.example.playlistmaker.Search.domain.model.Track
+
+interface TrackListClickListenerInterface {
+    fun onClick(track: Track) {}
+}
