+++ app/src/main/java/com/example/playlistmaker/Search/domain/api/NetworkClient.kt
@@ -1,0 +1,9 @@
+package com.example.playlistmaker.Search.domain.api
+
+import com.example.playlistmaker.Search.data.dto.NetworkResponse
+
+
+interface NetworkClient {
+    fun doRequest(dto: Any): NetworkResponse
+
+}
