+++ app/src/main/java/com/example/playlistmaker/Search/domain/api/SearchTrackInteractor.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.Search.domain.api
+
+import com.example.playlistmaker.Search.domain.model.Track
+
+interface SearchTrackInteractor {
+    fun searchTrack(expression: String, consumer: TrackConsumer)
+
+    interface TrackConsumer {
+        fun consume(foundTracks: List<Track>)
+    }
+}
