+++ app/src/main/java/com/example/playlistmaker/Search/domain/api/SearchTrackRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.Search.domain.api
+
+import com.example.playlistmaker.Search.domain.model.Track
+
+interface SearchTrackRepository {
+    fun searchTrack(expression: String): List<Track>
+}
