+++ app/src/main/java/com/example/playlistmaker/Search/domain/model/SearchConst.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.Search.domain.model
+
+object SearchConst {
+    const val SEARCH_TEXT = "SEARCH_TEXT"
+    const val MAX_SEARCH_COUNT = 10
+    const val ITUNES_BASE_URL = "https://itunes.apple.com"
+    const val SEARCH_DEBOUNCE_DELAY = 2000L
+}
