+++ app/src/main/java/com/example/playlistmaker/Search/domain/model/SearchStatus.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.Search.domain.model
+
+
+enum class SearchStatus {
+    PROGRESS_BAR,
+    SEARCH_HISTORY,
+    ERROR_EMPTY,
+    ERROR_BAD_CONNECTION,
+    SEARCH_RESULT,
+    NONE
+}
