--- app/src/main/java/com/example/playlistmaker/SearchResponse.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker
-
-data class SearchResponse(
-    val resultCount: Int,
-    val results: List<Track>
-)
-
