--- app/src/main/java/com/example/playlistmaker/TrackListClickListenerInterface.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker
-
-interface TrackListClickListenerInterface {
-    fun onClick(track: Track) {}
-}
